[수정] : 중복되는 id, 추상클래스로 이동
- AbstractEntity 클래스에 id 코드 추가
