+++ src/main/java/com/codesquad/qna/domain/User.java
@@ -25,6 +25,14 @@ public class User {
     @Column(nullable = false)
     private String email;
 
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
     public String getUserId() {
         return userId;
     }
@@ -55,14 +63,6 @@ public class User {
 
     public void setEmail(String email) {
         this.email = email;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
     }
 
     public void update(User updateUser) {
